add changes to scripts.js file in js folder

+++ js/scripts.js
@@ -3,6 +3,6 @@ var bodyMassIndex = function(weight, height) {
 };
 
 var height = parseInt(prompt("Enter your weight in kilograms:"));
-var weight = parseInt(prompt("Enter your height in meters squared:"));
+var weight = parseFloat(prompt("Enter your height in meters squared:"));
 var result = bodyMassIndex(weight, height);
 alert(result);
